scripts/mcap_tools/pyproject.toml
[project]
name = "mcap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcap-ros2-support>=0.5.5",
]
